pyproject.toml
[tool.poetry]
name = "polimpact"
version = "0.1.0"
description = "A long-term simulator of socio-political systems"
authors = ["Jonas Vestby <jonves@prio.org>"]
license = "MIT"
readme = "readme.md"

[tool.setuptools_scm]

[tool.poetry.dependencies]
python = "~3.10"
scikit-learn = "1.2.2"
numpy = "<1.24.0"
pandas = { version = "2.0.1", extras = ["performance", "output_formatting", "computation"] }
pandera = {extras = ["hypotheses", "strategies"], version = "^0.15.1"}
pillow = "^9.5.0"
cairosvg = "^2.7.0"
sklearn-xarray = "^0.4.0"
matplotlib = "^3.7.2"
bambi = "^0.12.0"
ipython = "^8.14.0"
ipykernel = "^6.25.1"
ipywidgets = "^8.1.0"
mlforecast = "^0.9.1"
hierarchicalforecast = "^0.3.0"
neuralforecast = "^1.6.2"
seaborn = "^0.12.2"
formulae = "^0.5.0"
hydra-core = "^1.3.2"
pydantic = "^2.3.0"
utilsforecast = "^0.0.12"
numbagg = "^0.6.0"
sphinxcontrib-mermaid = "^0.9.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[tool.poetry.group.docs.dependencies]
sphinx-inline-tabs = "^2023.4.21"
sphinx-copybutton = "^0.5.2"
furo = "^2023.7.26"
myst-parser = "^2.0.0"
nbsphinx = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
